[ADD] add tests for assets permissions

+++ app/Http/Controllers/Tenants/ForceDeleteAssetController.php
@@ -2,26 +2,27 @@
 
 namespace App\Http\Controllers\Tenants;
 
+use App\Helpers\ApiResponse;
 use Illuminate\Http\Request;
 use App\Models\Tenants\Asset;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 class ForceDeleteAssetController extends Controller
 {
-    public function forceDelete(int $assetId)
+    public function forceDelete(Asset $asset)
     {
-        $asset = Asset::onlyTrashed()->findOrFail($assetId);
+        if (Auth::user()->cannot('forceDelete', $asset))
+            abort(403);
+
         $asset->forceDelete();
 
-
-
-
         $tenantId = tenancy()->tenant->id;
-        $directory = "$tenantId/assets/$assetId/";
+        $directory = "$tenantId/assets/$asset->id/";
 
         Storage::disk('tenants')->deleteDirectory($directory);
 
-        return redirect()->route('tenant.assets.index');
+        return ApiResponse::success('', 'Asset deleted');
     }
 }
